+++ docs/changelog.rst
@@ -13,6 +13,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - hints kitten: Allow specifying :option:`kitty +kitten hints --program`
   multiple times to run multiple programs  (:iss:`1879`)
 
+- Add a :opt:`kitten_alias` option that can be used to alias kitten invocation
+  for brevity and to change kitten option defaults globally (:iss:`1879`)
+
 - Dont fail to start if running the shell to read the EDITOR env var fails
   (:iss:`1869`)